+++ src/ThumbTrait.php
@@ -14,6 +14,9 @@
 namespace Thumber;
 
 use Cake\Core\Configure;
+use Cake\Core\Plugin;
+use Cake\Filesystem\Folder;
+use Cake\Network\Exception\InternalErrorException;
 use Cake\Routing\Router;
 
 /**
@@ -83,5 +86,40 @@ trait ThumbTrait
     protected function getUrl($path, $full = true)
     {
         return Router::url(['_name' => 'thumb', base64_encode(basename($path))], $full);
+    }
+
+    /**
+     * Internal method to resolve a partial path, returning its full path
+     * @param string $path Partial path
+     * @return string
+     * @throws InternalErrorException
+     */
+    protected function resolveFilePath($path)
+    {
+        //Returns, if it's a remote file
+        if (isUrl($path)) {
+            return $path;
+        }
+
+        //If it a relative path, it can be a file from a plugin or a file
+        //  relative to `APP/webroot/img/`
+        if (!Folder::isAbsolute($path)) {
+            $pluginSplit = pluginSplit($path);
+
+            //Note that using `pluginSplit()` is not sufficient, because
+            //  `$path` may still contain a dot
+            if (!empty($pluginSplit[0]) && in_array($pluginSplit[0], Plugin::loaded())) {
+                $path = Plugin::path($pluginSplit[0]) . 'webroot' . DS . 'img' . DS . $pluginSplit[1];
+            } else {
+                $path = WWW_ROOT . 'img' . DS . $path;
+            }
+        }
+
+        //Checks if is readable
+        if (!is_readable($path)) {
+            throw new InternalErrorException(__d('thumber', 'File `{0}` not readable', rtr($path)));
+        }
+
+        return $path;
     }
 }